+++ example/lib/counter/counter_screen_animation.dart
@@ -48,7 +48,7 @@ class _CounterScreenWithAnimationState extends State<CounterScreenWithAnimation>
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
               Text(Cubes.getString('description_counter')),
-              cube!.count.build<int>((value) {
+              cube.count.build<int>((value) {
                 return Text(value.toString());
               }),
             ],
@@ -56,7 +56,7 @@ class _CounterScreenWithAnimationState extends State<CounterScreenWithAnimation>
         ),
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: cube!.increment,
+        onPressed: cube.increment,
         tooltip: 'Increment',
         child: Icon(Icons.add),
       ), // This trailing comma makes auto-formatting nicer for build methods.